moved dynamic data test to dom elements tests
